Use 64 bit types for memory allocation totals even on 32 bit

+++ src/boehm_gc.c
@@ -50,7 +50,7 @@ enum { NTYPES = 256 };
 
 TNumInfoBags InfoBags[NTYPES];
 
-UInt SizeAllBags;
+UInt8 SizeAllBags;
 
 static inline Bag * DATA(BagHeader * bag)
 {
